requirements.txt
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
Werkzeug==2.1.1
zipp==3.8.0
